Move retrofit to data layer.
