+++ app/src/main/java/com/inu/cafeteria/injection/RetrofitFactory.kt
@@ -23,7 +23,7 @@ import android.content.Context
 import com.franmontiel.persistentcookiejar.PersistentCookieJar
 import com.franmontiel.persistentcookiejar.cache.SetCookieCache
 import com.franmontiel.persistentcookiejar.persistence.SharedPrefsCookiePersistor
-import com.inu.cafeteria.service.CafeteriaNetworkService
+import com.inu.cafeteria.retrofit.CafeteriaNetworkService
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
