+++ data/src/main/java/com/inu/cafeteria/retrofit/CafeteriaNetworkService.kt
@@ -17,9 +17,28 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.inu.cafeteria.service
+/**
+ * This file is part of INU Cafeteria.
+ *
+ * Copyright (C) 2020 INU Global App Center <potados99@gmail.com>
+ *
+ * INU Cafeteria is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * INU Cafeteria is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
 
-import com.inu.cafeteria.model.scheme.*
+package com.inu.cafeteria.retrofit
+
+import com.inu.cafeteria.retrofit.scheme.*
 import retrofit2.Call
 import retrofit2.http.*
 
@@ -29,8 +48,13 @@ import retrofit2.http.*
 
 interface CafeteriaNetworkService {
     @GET("/notices")
-    fun getNotice(): Call<Notice>
+    fun getAllNotices(): Call<List<NoticeResult>>
 
+    @GET("/notices/latest")
+    fun getLatestNotice(): Call<NoticeResult>
+
+    @GET("/shouldIUpdate")
+    fun shouldIUpdate(@Query("os") os: String, @Query("version") version: String): Call<Boolean>
 
     @GET("/cafeteria")
     fun getCafeteria(): Call<List<CafeteriaResult>>
