+++ data/src/main/java/com/inu/cafeteria/retrofit/scheme/CornerResult.kt
@@ -17,12 +17,14 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.inu.cafeteria.model.scheme
+package com.inu.cafeteria.retrofit.scheme
 
-/**
- * Logout request scheme.
- */
+import com.google.gson.annotations.SerializedName
 
-data class LogoutParams(
-    val token: String   // The token we got on login
+data class CornerResult(
+    val id: Int,
+    val name: String,
+    @SerializedName("display-name") val displayName: String,
+    @SerializedName("available-at") val availableAt: Int,
+    @SerializedName("cafeteria-id") val cafeteriaId: Int
 )
